+++ Makefile.in
@@ -2264,7 +2264,7 @@ COND_PLATFORM_WIN32_1_GTK_PLATFORM_HDR =  \
 COND_TOOLKIT_MSW_WEBVIEW_HDR_PLATFORM =  \
 	wx/msw/webviewhistoryitem_ie.h \
 	wx/msw/webview_ie.h \
-	wx/msw/webview_edgec.h
+	wx/msw/webview_edge.h
 @COND_TOOLKIT_MSW@WEBVIEW_HDR_PLATFORM = $(COND_TOOLKIT_MSW_WEBVIEW_HDR_PLATFORM)
 @COND_TOOLKIT_COCOA@OPENGL_HDR_PLATFORM = wx/cocoa/glcanvas.h
 @COND_TOOLKIT_GTK_TOOLKIT_VERSION_@OPENGL_HDR_PLATFORM \
@@ -6137,7 +6137,7 @@ COND_TOOLKIT_X11___LOWLEVEL_SRC_OBJECTS_1 =  \
 @COND_TOOLKIT_GTK@__WEBVIEW_SRC_PLATFORM_OBJECTS = \
 @COND_TOOLKIT_GTK@	monodll_gtk_webview_webkit.o monodll_webview_webkit2.o
 @COND_TOOLKIT_MSW@__WEBVIEW_SRC_PLATFORM_OBJECTS = \
-@COND_TOOLKIT_MSW@	monodll_webview_ie.o monodll_webview_edgec.o
+@COND_TOOLKIT_MSW@	monodll_webview_ie.o monodll_webview_edge.o
 @COND_TOOLKIT_GTK_TOOLKIT_VERSION_2@__AUI_GTK_SRC_OBJECTS \
 @COND_TOOLKIT_GTK_TOOLKIT_VERSION_2@	= monodll_tabartgtk.o
 @COND_TOOLKIT_MSW@__AUI_PLATFORM_SRC_OBJECTS = \
@@ -8127,7 +8127,7 @@ COND_TOOLKIT_X11___LOWLEVEL_SRC_OBJECTS_3 =  \
 @COND_TOOLKIT_GTK@__WEBVIEW_SRC_PLATFORM_OBJECTS_1 = \
 @COND_TOOLKIT_GTK@	monolib_gtk_webview_webkit.o monolib_webview_webkit2.o
 @COND_TOOLKIT_MSW@__WEBVIEW_SRC_PLATFORM_OBJECTS_1 = \
-@COND_TOOLKIT_MSW@	monolib_webview_ie.o monolib_webview_edgec.o
+@COND_TOOLKIT_MSW@	monolib_webview_ie.o monolib_webview_edge.o
 @COND_TOOLKIT_GTK_TOOLKIT_VERSION_2@__AUI_GTK_SRC_OBJECTS_1 \
 @COND_TOOLKIT_GTK_TOOLKIT_VERSION_2@	= monolib_tabartgtk.o
 @COND_TOOLKIT_MSW@__AUI_PLATFORM_SRC_OBJECTS_1 = \
@@ -12336,7 +12336,7 @@ COND_USE_SOVERSOLARIS_1___webviewdll___so_symlinks_uninst_cmd = rm -f \
 @COND_TOOLKIT_GTK@	webviewdll_gtk_webview_webkit.o \
 @COND_TOOLKIT_GTK@	webviewdll_webview_webkit2.o
 @COND_TOOLKIT_MSW@__WEBVIEW_SRC_PLATFORM_OBJECTS_2 = \
-@COND_TOOLKIT_MSW@	webviewdll_webview_ie.o webviewdll_webview_edgec.o
+@COND_TOOLKIT_MSW@	webviewdll_webview_ie.o webviewdll_webview_edge.o
 @COND_TOOLKIT_MSW@__webview_additional_include_p_1 = \
 @COND_TOOLKIT_MSW@	--include-dir \
 @COND_TOOLKIT_MSW@	$(top_srcdir)/3rdparty/webview2/build/native/include
@@ -12360,7 +12360,7 @@ COND_MONOLITHIC_0_SHARED_0_USE_GUI_1_USE_WEBVIEW_1___webviewlib___depname = \
 @COND_TOOLKIT_GTK@	webviewlib_gtk_webview_webkit.o \
 @COND_TOOLKIT_GTK@	webviewlib_webview_webkit2.o
 @COND_TOOLKIT_MSW@__WEBVIEW_SRC_PLATFORM_OBJECTS_3 = \
-@COND_TOOLKIT_MSW@	webviewlib_webview_ie.o webviewlib_webview_edgec.o
+@COND_TOOLKIT_MSW@	webviewlib_webview_ie.o webviewlib_webview_edge.o
 @COND_SHARED_1@____wxwebview_namedll_DEP = $(__webviewdll___depname)
 @COND_SHARED_0@____wxwebview_namelib_DEP = $(__webviewlib___depname)
 COND_MONOLITHIC_0_SHARED_1_USE_GUI_1_USE_QA_1___qadll___depname = \
@@ -17119,8 +17119,8 @@ monodll_htmllbox.o: $(srcdir)/src/generic/htmllbox.cpp $(MONODLL_ODEP)
 monodll_webview_ie.o: $(srcdir)/src/msw/webview_ie.cpp $(MONODLL_ODEP)
 	$(CXXC) -c -o $@ $(MONODLL_CXXFLAGS) $(srcdir)/src/msw/webview_ie.cpp
 
-monodll_webview_edgec.o: $(srcdir)/src/msw/webview_edgec.cpp $(MONODLL_ODEP)
-	$(CXXC) -c -o $@ $(MONODLL_CXXFLAGS) $(srcdir)/src/msw/webview_edgec.cpp
+monodll_webview_edge.o: $(srcdir)/src/msw/webview_edge.cpp $(MONODLL_ODEP)
+	$(CXXC) -c -o $@ $(MONODLL_CXXFLAGS) $(srcdir)/src/msw/webview_edge.cpp
 
 monodll_gtk_webview_webkit.o: $(srcdir)/src/gtk/webview_webkit.cpp $(MONODLL_ODEP)
 	$(CXXC) -c -o $@ $(MONODLL_CXXFLAGS) $(srcdir)/src/gtk/webview_webkit.cpp
@@ -22399,8 +22399,8 @@ monolib_htmllbox.o: $(srcdir)/src/generic/htmllbox.cpp $(MONOLIB_ODEP)
 monolib_webview_ie.o: $(srcdir)/src/msw/webview_ie.cpp $(MONOLIB_ODEP)
 	$(CXXC) -c -o $@ $(MONOLIB_CXXFLAGS) $(srcdir)/src/msw/webview_ie.cpp
 
-monolib_webview_edgec.o: $(srcdir)/src/msw/webview_edgec.cpp $(MONOLIB_ODEP)
-	$(CXXC) -c -o $@ $(MONOLIB_CXXFLAGS) $(srcdir)/src/msw/webview_edgec.cpp
+monolib_webview_edge.o: $(srcdir)/src/msw/webview_edge.cpp $(MONOLIB_ODEP)
+	$(CXXC) -c -o $@ $(MONOLIB_CXXFLAGS) $(srcdir)/src/msw/webview_edge.cpp
 
 monolib_gtk_webview_webkit.o: $(srcdir)/src/gtk/webview_webkit.cpp $(MONOLIB_ODEP)
 	$(CXXC) -c -o $@ $(MONOLIB_CXXFLAGS) $(srcdir)/src/gtk/webview_webkit.cpp
@@ -35983,8 +35983,8 @@ htmllib_htmllbox.o: $(srcdir)/src/generic/htmllbox.cpp $(HTMLLIB_ODEP)
 webviewdll_webview_ie.o: $(srcdir)/src/msw/webview_ie.cpp $(WEBVIEWDLL_ODEP)
 	$(CXXC) -c -o $@ $(WEBVIEWDLL_CXXFLAGS) $(srcdir)/src/msw/webview_ie.cpp
 
-webviewdll_webview_edgec.o: $(srcdir)/src/msw/webview_edgec.cpp $(WEBVIEWDLL_ODEP)
-	$(CXXC) -c -o $@ $(WEBVIEWDLL_CXXFLAGS) $(srcdir)/src/msw/webview_edgec.cpp
+webviewdll_webview_edge.o: $(srcdir)/src/msw/webview_edge.cpp $(WEBVIEWDLL_ODEP)
+	$(CXXC) -c -o $@ $(WEBVIEWDLL_CXXFLAGS) $(srcdir)/src/msw/webview_edge.cpp
 
 webviewdll_gtk_webview_webkit.o: $(srcdir)/src/gtk/webview_webkit.cpp $(WEBVIEWDLL_ODEP)
 	$(CXXC) -c -o $@ $(WEBVIEWDLL_CXXFLAGS) $(srcdir)/src/gtk/webview_webkit.cpp
@@ -36010,8 +36010,8 @@ webviewdll_version_rc.o: $(srcdir)/src/msw/version.rc $(WEBVIEWDLL_ODEP)
 webviewlib_webview_ie.o: $(srcdir)/src/msw/webview_ie.cpp $(WEBVIEWLIB_ODEP)
 	$(CXXC) -c -o $@ $(WEBVIEWLIB_CXXFLAGS) $(srcdir)/src/msw/webview_ie.cpp
 
-webviewlib_webview_edgec.o: $(srcdir)/src/msw/webview_edgec.cpp $(WEBVIEWLIB_ODEP)
-	$(CXXC) -c -o $@ $(WEBVIEWLIB_CXXFLAGS) $(srcdir)/src/msw/webview_edgec.cpp
+webviewlib_webview_edge.o: $(srcdir)/src/msw/webview_edge.cpp $(WEBVIEWLIB_ODEP)
+	$(CXXC) -c -o $@ $(WEBVIEWLIB_CXXFLAGS) $(srcdir)/src/msw/webview_edge.cpp
 
 webviewlib_gtk_webview_webkit.o: $(srcdir)/src/gtk/webview_webkit.cpp $(WEBVIEWLIB_ODEP)
 	$(CXXC) -c -o $@ $(WEBVIEWLIB_CXXFLAGS) $(srcdir)/src/gtk/webview_webkit.cpp
